Meshing works now, just need to keep advancing!

+++ Bounce/Classes/[Bounce]/IndexTriangleList.swift
@@ -1,22 +1,20 @@
 //
 //  IndexTriangleList.swift
-//  Bounce
 //
 //  Created by Raptis, Nicholas on 9/2/16.
-//  Copyright © 2016 Darkswarm LLC. All rights reserved.
 //
 
 import Foundation
 
 struct IndexTriangle {
-    var x1:Int16 = 0
-    var y1:Int16 = 0
+    var x1:Int = 0
+    var y1:Int = 0
     
-    var x2:Int16 = 0
-    var y2:Int16 = 0
+    var x2:Int = 0
+    var y2:Int = 0
     
-    var x3:Int16 = 0
-    var y3:Int16 = 0
+    var x3:Int = 0
+    var y3:Int = 0
 }
 
 class IndexTriangleList {
@@ -45,12 +43,12 @@ class IndexTriangleList {
             }
         }
         
-        data[_count].x1 = Int16(x1)
-        data[_count].y1 = Int16(y1)
-        data[_count].x2 = Int16(x2)
-        data[_count].y2 = Int16(y2)
-        data[_count].x3 = Int16(x3)
-        data[_count].y3 = Int16(y3)
+        data[_count].x1 = Int(x1)
+        data[_count].y1 = Int(y1)
+        data[_count].x2 = Int(x2)
+        data[_count].y2 = Int(y2)
+        data[_count].x3 = Int(x3)
+        data[_count].y3 = Int(y3)
         
         _count += 1
     }
